Add protocol for signalling consumption from consumer to producer
IPC message can optionally transfer buffers from the producer to
the consumer. These buffers are not copied while the message is in
transit. Consequently, the consumer must inform the producer when
it has finished processing the buffer.

This patch adds a monitor variable that the producer uses to wait
for the competion of the consumers message handling. Once done, the
consumer informs the producer to continue.

The reply can also return a buffer back to the consumer. The lifetime
of this buffer is managed by a higher-level protocol.